requirements.txt
Flask==3.0.0
opencv-python-headless==4.9.0.80
PyYAML==6.0.1
ultralytics==8.1.0
Werkzeug==3.0.1
Pillow==10.2.0
gunicorn==21.2.0
